Commented to Avoid Compile Time Errors
We found a bug. In the last release we changed .NET’s out/ref params emulation. Earlier it was an array, now it is Ref<> class. Zz95<Object> really is Ref<Object>. We should make the class public. We just added an JIRA issue for this (WORDSJAVA-1756).

Anyway, it was big changes of Samples and Examples code on .NET baseline. We are planing to port all the changes in the next 18.4 release.